fix: Use a cache again, to handle if any recently used video ids are called from some unknown use case
